+++ CHANGELOG.md
@@ -1,6 +1,30 @@
 # Release Notes
 
-## [Unreleased](https://github.com/laravel/mcp/compare/v0.3.0...main)
+## [Unreleased](https://github.com/laravel/mcp/compare/v0.3.3...main)
+
+## [v0.3.3](https://github.com/laravel/mcp/compare/v0.3.2...v0.3.3) - 2025-11-11
+
+* Add MCP service provider to testbench config by [@zacksmash](https://github.com/zacksmash) in https://github.com/laravel/mcp/pull/100
+* Fix client_name rename in oauth registrar by [@mikebouwmans](https://github.com/mikebouwmans) in https://github.com/laravel/mcp/pull/104
+* fix: allow multi-segment issuer paths by [@isaac-bowen](https://github.com/isaac-bowen) in https://github.com/laravel/mcp/pull/105
+
+## [v0.3.2](https://github.com/laravel/mcp/compare/v0.3.1...v0.3.2) - 2025-10-29
+
+### What's Changed
+
+* Ensure the property field exists in tool input schemas by [@pushpak1300](https://github.com/pushpak1300) in https://github.com/laravel/mcp/pull/97
+
+### New Contributors
+
+* [@pushpak1300](https://github.com/pushpak1300) made their first contribution in https://github.com/laravel/mcp/pull/97
+
+**Full Changelog**: https://github.com/laravel/mcp/compare/v0.3.1...v0.3.2
+
+## [v0.3.1](https://github.com/laravel/mcp/compare/v0.3.0...v0.3.1) - 2025-10-24
+
+* refactor: move to first class callable by [@ashleyhindle](https://github.com/ashleyhindle) in https://github.com/laravel/mcp/pull/94
+* Cast `client_id` to string in JSON response by [@mostafa-rz](https://github.com/mostafa-rz) in https://github.com/laravel/mcp/pull/93
+* Feature: adds security to the OAuth registration endpoint by [@jsandfordhughescoop](https://github.com/jsandfordhughescoop) in https://github.com/laravel/mcp/pull/87
 
 ## [v0.3.0](https://github.com/laravel/mcp/compare/v0.2.1...v0.3.0) - 2025-10-07